Add WPCheckout wrapper component
