+++ packages/composite-checkout-wpcom/src/index.js
@@ -1,10 +1,10 @@
 /**
  * Internal dependencies
  */
-import { WPCOMCheckout } from './components/checkout';
+import { WPCheckoutWrapper } from './components/checkout';
 import { makeShoppingCartHook } from './hooks/cart-manager';
 import { mockCartEndpoint } from './mock/cart-endpoint';
 import { mockPayPalExpressRequest } from './mock/paypal-payment-method';
 
 // Re-export the public API
-export { WPCOMCheckout, makeShoppingCartHook, mockCartEndpoint, mockPayPalExpressRequest };
+export { WPCheckoutWrapper, makeShoppingCartHook, mockCartEndpoint, mockPayPalExpressRequest };
